move acc to first value
