+++ Cargo.toml
@@ -74,11 +74,11 @@ chrono = { version = "0.4.34", default-features = false }
 ctor = "0.2.0"
 dashmap = "5.4.0"
 datafusion = { path = "datafusion/core", version = "37.0.0", default-features = false }
+datafusion-aggregate-functions = { path = "datafusion/aggregate-functions", version = "37.0.0" }
 datafusion-common = { path = "datafusion/common", version = "37.0.0", default-features = false }
 datafusion-common-runtime = { path = "datafusion/common-runtime", version = "37.0.0" }
 datafusion-execution = { path = "datafusion/execution", version = "37.0.0" }
 datafusion-expr = { path = "datafusion/expr", version = "37.0.0" }
-datafusion-aggregate-functions = { path = "datafusion/aggregate-functions", version = "37.0.0" }
 datafusion-functions = { path = "datafusion/functions", version = "37.0.0" }
 datafusion-functions-array = { path = "datafusion/functions-array", version = "37.0.0" }
 datafusion-optimizer = { path = "datafusion/optimizer", version = "37.0.0", default-features = false }
