+++ datafusion/expr/src/utils.rs
@@ -1240,6 +1240,12 @@ pub fn merge_schema(inputs: Vec<&LogicalPlan>) -> DFSchema {
     }
 }
 
+/// Construct state name. State is the intermidiate state of the aggregate function.
+/// TODO: Remove duplicated function in physical-expr
+pub fn format_state_name(name: &str, state_name: &str) -> String {
+    format!("{name}[{state_name}]")
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
